requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
pydantic
python-dotenv
pytest
httpx
pytest-asyncio
passlib[bcrypt]
python-jose[cryptography]
greenlet